+++ packages/spacecat-shared-data-access/test/it/util/seed.js
@@ -26,6 +26,8 @@ const resetDatabase = async () => {
     TEST_DA_CONFIG.tableNameExperiments,
     TEST_DA_CONFIG.tableNameImportJobs,
     TEST_DA_CONFIG.tableNameImportUrls,
+    TEST_DA_CONFIG.tableNameScrapeJobs,
+    TEST_DA_CONFIG.tableNameScrapeUrls,
     TEST_DA_CONFIG.tableNameKeyEvents,
     TEST_DA_CONFIG.tableNameLatestAudits,
     TEST_DA_CONFIG.tableNameOrganizations,